Refactor: move game object classes and card data sets to sub folders
